3. database/requirements.txt
supabase>=2.0.3
python-dotenv>=1.0.0
httpx>=0.24.0,<0.25.0